fix: master branch ci failed
指定qmake参数，遇到错误退出

Log: CI错误修复

+++ tests/test-recoverage-qmake.sh
@@ -1,22 +1,23 @@
 #!/bin/bash
 
+set -e
 BUILD_DIR=build
 REPORT_DIR=report
 #EXTRACT_ARGS="src"
-cd ../
-rm -rf $BUILD_DIR
-mkdir $BUILD_DIR
-cd $BUILD_DIR
-qmake .. CONFIG+=debug
-make -j$(nproc)
-cd ../tests/
+#cd ../
+#rm -rf $BUILD_DIR
+#mkdir $BUILD_DIR
+#cd $BUILD_DIR
+#qmake ../ CONFIG+=debug BASED_DTK_DIR=based-dtk DTK_VERSION=5.5
+#make -j$(nproc)
+#cd ../tests/
 
 rm -rf $BUILD_DIR
 mkdir $BUILD_DIR
 cd $BUILD_DIR
-qmake ../ CONFIG+=debug
+qmake ../ CONFIG+=debug BASED_DTK_DIR=based-dtk DTK_VERSION=5.5
 export ASAN_OPTIONS=halt_on_error=0
-TESTARGS="--gtest_output=xml:dde_test_report_qt5integration.xml" make check -j$(nproc)
+TESTARGS="--gtest_output=xml:dde_test_report_qt5integration.xml" make qmake_all check -j$(nproc)
 
 lcov -d ./ -c -o coverage_all.info
 #lcov --extract coverage_all.info $EXTRACT_ARGS --output-file coverage.info
@@ -24,7 +25,7 @@ lcov --remove coverage_all.info "*/tests/*" "*/usr/include*" "*build/src*" --out
 cd ..
 genhtml -o $REPORT_DIR $BUILD_DIR/coverage.info
 
-mv ./build/asan.log* ./build/asan_qt5integration.log
+test -e ./build/asan.log* && mv ./build/asan.log* ./build/asan_qt5integration.log || touch ./build/asan.log
 
 #rm -rf $BUILD_DIR
 #rm -rf ../$BUILD_DIR
